requirements.txt
mlflow==2.13.0
scikit-learn==1.4.2
pandas==2.2.2
dvc[gs]==3.50.1
pathspec==0.11.2
pytest==8.2.0
fastapi==0.111.0
uvicorn==0.29.0
joblib==1.4.2
google-cloud-storage==2.16.0
pyyaml==6.0.1
xgboost==2.0.3
setuptools